update form init format
