+++ amis_py/components/page.py
@@ -1,0 +1,4 @@
+from amis_py.components.base import Prop
+
+
+class PageProperties:
